fix: garbage collect not implemented error

+++ src/model/cloud-runner/providers/docker/index.ts
@@ -144,7 +144,12 @@ cp -a ${sharedFolder}. /github/workspace/cloud-runner-cache/
       true,
     );
 
-    if (!this.buildParameters?.retainWorkspace) {
+    if (
+      !this.buildParameters?.retainWorkspace &&
+      fs.existsSync(
+        `${workspace}/cloud-runner-cache/${path.basename(CloudRunnerFolders.uniqueCloudRunnerJobFolderAbsolute)}`,
+      )
+    ) {
       await CloudRunnerSystem.Run(
         `rm -r ${workspace}/cloud-runner-cache/${path.basename(CloudRunnerFolders.uniqueCloudRunnerJobFolderAbsolute)}`,
       );
